loan_system_backend/requirements.txt
aiofiles==25.1.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==5.0.0
certifi==2026.1.4
greenlet==3.3.1
h11==0.16.0
html5tagger==1.3.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
multidict==6.7.1
packaging==26.0
passlib==1.7.4
PyJWT==2.11.0
python-dotenv==1.2.1
sanic==25.12.0
Sanic-Cors==2.2.0
sanic-routing==23.12.0
setuptools==82.0.0
SQLAlchemy==2.0.46
tracerite==2.3.1
typing_extensions==4.15.0
websockets==16.0